rename testimport to testall
