--- Tests/hardcodeddata.h
@@ -1,19 +1,0 @@
-#ifndef HARDCODEDDATA_H
-#define HARDCODEDDATA_H
-
-#include "json/json.h"
-
-//The expected default debugcsv
-const char * ___debugcsvtxt = R"quotePower( {	"columns" : 	[		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1",				"2",				"3",				"4",				"5",				"6",				"7",				"8",				"9",				"10",				"11",				"12",				"13",				"14",				"15",				"16",				"17",				"18",				"19",				"20",				"21",				"22",				"23",				"24",				"25",				"26",				"27",				"28",				"29",				"30",				"31",				"32",				"33",				"34",				"35",				"36",				"37",				"38",				"39",				"40",				"41",				"42",				"43",				"44",				"45",				"46",				"47",				"48",				"49",				"50",				"51",				"52",				"53",				"54",				"55",				"56",				"57",				"58",				"59",				"60",				"61",				"62",				"63",				"64",				"65",				"66",				"67",				"68",				"69",				"70",				"71",				"72",				"73",				"74",				"75",				"76",				"77",				"78",				"79",				"80",				"81",				"82",				"83",				"84",				"85",				"86",				"87",				"88",				"89",				"90",				"91",				"92",				"93",				"94",				"95",				"96",				"97",				"98",				"99",				"100" 			], 			"description" : "",			"name" : "V1",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"0.99122228",				"-1.953344972",				"1.576301642",				"0.947933415",				"-0.532552496",				"-1.363597632",				"-0.654979378",				"-0.751445953",				"0.310181524",				"-0.675270732",				"0.364848251",				"0.525950287",				"-0.954597608",				"-0.399043745",				"-2.336742886",				"-0.632883125",				"0.588682079",				"0.391750556",				"-1.264009159",				"1.058565048",				"-0.850244619",				"-0.359553637",				"0.108920494",				"0.01096381",				"2.739343353",				"-0.59740969",				"-0.470163224",				"-1.447375559",				"0.505173357",				"0.307537423",				"-1.395443011",				"-0.367996707",				"0.793185108",				"0.596201293",				"-0.556843301",				"0.644058891",				"-0.242813433",				"1.410001008",				"-0.342746069",				"-0.618288772",				"-0.692689912",				"-1.415363942",				"-0.786160472",				"-0.840984252",				"-0.999014305",				"1.007309698",				"-0.611664173",				"0.46937728",				"-3.023963827",				"0.166778121",				"-0.658287055",				"-1.649937891",				"0.757371551",				"-0.412828649",				"3.356094448",				"-0.656919662",				"0.260454418",				"0.092206944",				"-0.0767995",				"0.284253602",				"0.972104397",				"-1.375547175",				"-1.627592736",				"2.179421126",				"0.376149862",				"-1.144123942",				"1.889051803",				"0.176474657",				"-0.667823773",				"0.570675169",				"-0.881899561",				"-0.801894546",				"-0.120997408",				"-0.937671609",				"-2.15616121",				"0.441266475",				"-0.712944476",				"0.503236412",				"-1.352650303",				"0.259356367",				"-0.934349142",				"0.518141224",				"2.958797116",				"-0.118805371",				"-0.805832374",				"-0.096001487",				"0.239571257",				"-0.199426499",				"-0.805996397",				"-0.707972381",				"-1.625143884",				"-0.648908102",				"-0.051036508",				"-0.433036966",				"-0.398710373",				"0.337436519",				"-0.434493855",				"-0.443990603",				"-1.030775205",				"0.544538215" 			], 			"description" : "",			"name" : "contNormal",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"0.780975715",				"1.661185538",				"1.176677112",				"1.331417271",				"1.18426484",				"2.478443291",				"1.529750756",				"0.348246238",				"0.37865767",				"2.777961973",				"3.977031334",				"2.656664043",				"0.441790033",				"0.678039819",				"2.530188739",				"3.306006111",				"1.900242487",				"1.761105829",				"2.29457509",				"0.942826432",				"2.833575102",				"1.839209197",				"5.078108173",				"1.383094211",				"0.527073852",				"4.618063242",				"5.157894077",				"2.363390504",				"2.838679609",				"0.660057008",				"3.930940809",				"2.230799959",				"1.160636689",				"0.197333889",				"0.676660939",				"4.975727839",				"2.77466792",				"1.532106738",				"3.959582989",				"0.991611379",				"0.722313192",				"0.614345924",				"0.483241669",				"2.359102511",				"1.750826816",				"4.321780883",				"1.195401638",				"2.984005893",				"0.410306148",				"0.302480929",				"0.494589392",				"0.715926748",				"1.281209304",				"2.775624599",				"1.644354059",				"2.248620658",				"4.279672814",				"0.849437567",				"1.68140103",				"2.563488535",				"1.255015115",				"1.333092118",				"1.478328803",				"2.674861547",				"0.432771654",				"2.692449102",				"0.862784728",				"3.999781721",				"4.661099397",				"4.949013358",				"1.792365981",				"0.683131832",				"0.343003988",				"2.029515062",				"2.231032155",				"1.800714959",				"1.294544523",				"0.934617004",				"8.768291675",				"3.491491238",				"2.403076495",				"1.711351491",				"0.781081083",				"1.125291522",				"0.43534484",				"2.236351222",				"0.636380241",				"5.869967659",				"1.268093974",				"1.710853925",				"1.78856175",				"1.314785303",				"2.035091245",				"3.096990053",				"0.820566944",				"5.539742862",				"0.540298922",				"0.479334111",				"3.015264007",				"1.236351257" 			], 			"description" : "",			"name" : "contGamma",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1",				"0",				"0",				"1",				"1",				"1",				"0",				"0",				"0",				"0",				"0",				"1",				"0",				"0",				"0",				"0",				"1",				"0",				"0",				"0",				"0",				"1",				"0",				"0",				"0",				"0",				"0",				"1",				"1",				"0",				"0",				"0",				"1",				"0",				"0",				"1",				"1",				"1",				"1",				"0",				"0",				"1",				"0",				"0",				"1",				"0",				"1",				"1",				"1",				"0",				"1",				"1",				"0",				"1",				"0",				"1",				"0",				"0",				"0",				"0",				"0",				"0",				"1",				"1",				"1",				"1",				"1",				"0",				"0",				"0",				"1",				"0",				"1",				"0",				"0",				"0",				"1",				"0",				"0",				"0",				"0",				"0",				"0",				"1",				"1",				"1",				"0",				"1",				"1",				"0",				"0",				"1",				"1",				"1",				"1",				"0",				"1",				"0",				"1",				"0" 			], 			"description" : "",			"name" : "contBinom",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1.7e-20",				"18858693.53",				"1.37e-24",				"1.02e+27",				"2.98e-26",				"1.9e+33",				"2.96e+29",				"1.01e+20",				"1.56e+35",				"1.74e-49",				"3.29e-21",				"8.89e-23",				"3.38e-27",				"5606.730444",				"3427.815043",				"2.04e+25",				"1.25e-32",				"10641603.08",				"0.042666202",				"0.008650772",				"2.26e-17",				"7.84e+29",				"2043267.37",				"204310864.6",				"5.37e-12",				"85379971.27",				"8.75e-07",				"0.503955858",				"1.2e+22",				"6.18e-06",				"0.002978766",				"28940.10745",				"1.61e-07",				"8.58e+33",				"1.67e-13",				"4.17e-07",				"5.22e+17",				"1.19e+23",				"5.37e+36",				"6.68e-23",				"5.87e-12",				"1.16e-26",				"527.1268561",				"48530.464",				"5.54e-13",				"6.2e-06",				"9.36e-22",				"3.44e+57",				"1.22e-17",				"9.71e-10",				"11470440.29",				"3.72e-16",				"1.65e-08",				"4.1e-05",				"6.13e-06",				"8e+15",				"0.037113115",				"3.03e-11",				"4.61e-22",				"6.71e-18",				"0.02922171",				"7.25e-26",				"9.81e+46",				"1.9e+12",				"1.82e-07",				"0.009103734",				"2.67e+47",				"1.11e-07",				"29946.53013",				"2.42e-16",				"1.84e+40",				"1.33e-21",				"3.42e-07",				"4.54e-08",				"2.07e+37",				"2.28e-11",				"1.18e-13",				"2.99e+49",				"1.22e-31",				"340227.7108",				"1.85e-20",				"7259386.785",				"6.71e+13",				"1.36e-15",				"1.12e-15",				"1.33e+35",				"4.56e+11",				"0.404686533",				"33.46701836",				"0.000189243",				"5.2e-34",				"1.43e-12",				"281634780.2",				"3.261864245e+10",				"5.5e-22",				"4.21e-21",				"6e-26",				"1.22e-06",				"3.61e+35",				"1.61e+34" 			], 			"description" : "",			"name" : "contExpon",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"8.38e+98",				"3.87e+99",				"7.02e+99",				"2.26e+98",				"5.76e+99",				"-4.65e+99",				"-6.49e+99",				"1.57e+99",				"5.14e+99",				"-3.52e+99",				"-7.75e+98",				"7.21e+99",				"-2.24e+99",				"-6.56e+99",				"8.55e+99",				"2.8e+98",				"-6.29e+99",				"-1.85e+99",				"-2.42e+99",				"-2.29e+99",				"5.73e+99",				"-7.1e+99",				"1.82e+99",				"-1.43e+99",				"-5.25e+99",				"-8.03e+99",				"3.15e+99",				"8.72e+99",				"-4.66e+99",				"-5.21e+99",				"-1.8e+99",				"-5.99e+99",				"8.93e+99",				"6.23e+99",				"2.24e+99",				"2.1e+98",				"7.75e+99",				"3.74e+99",				"2.51e+99",				"1.53e+99",				"-3.32e+99",				"-7.78e+99",				"-6.47e+99",				"5.92e+99",				"-6.9e+98",				"3.4e+99",				"8.46e+99",				"7.81e+99",				"-5.17e+99",				"-2.94e+99",				"8.27e+99",				"-3.59e+99",				"1.42e+99",				"1.88e+99",				"7.64e+99",				"-5.38e+99",				"-2.17e+99",				"-5.74e+99",				"5.71e+99",				"-5.49e+99",				"4.57e+99",				"1.95e+99",				"4.34e+98",				"-8e+99",				"6.5e+99",				"-5.2e+99",				"-8.71e+99",				"-3.6e+99",				"5.97e+99",				"4.46e+99",				"-3.25e+99",				"1.89e+99",				"7.65e+99",				"-8.96e+99",				"-5.84e+99",				"-7.03e+99",				"-2.52e+99",				"6.34e+99",				"4.6e+99",				"2.84e+99",				"-1.34e+99",				"-7.76e+99",				"4.62e+99",				"-4.08e+99",				"-6.32e+99",				"-2.17e+99",				"5.62e+99",				"-7.33e+99",				"-1.6e+99",				"-2.9e+98",				"7.42e+99",				"1.89e+99",				"-8.89e+99",				"-1.36e+99",				"2.57e+99",				"5.97e+98",				"5.77e+98",				"-3.79e+99",				"6.66e+98",				"2.83e+99" 			], 			"description" : "",			"name" : "contWide",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"-9.96e-100",				"-1.27e-100",				"-8.78e-100",				"-9.53e-100",				"8.33e-100",				"-8.3e-100",				"1.28e-100",				"7.53e-100",				"5.5e-100",				"-9.97e-100",				"-3.43e-100",				"-7.73e-100",				"2.59e-100",				"-5.86e-100",				"-7.26e-100",				"-3.51e-100",				"3.96e-100",				"2.61e-100",				"-6.76e-100",				"-1.1e-100",				"8.56e-100",				"-1.42e-100",				"5.85e-100",				"6.65e-100",				"-6.96e-100",				"-4.88e-100",				"3.83e-100",				"1.01e-100",				"9.31e-101",				"2.63e-100",				"1.21e-100",				"7.31e-100",				"7.23e-101",				"5.62e-100",				"2.77e-101",				"4.68e-101",				"7.61e-100",				"-2.98e-100",				"3.41e-100",				"4.91e-100",				"-1.72e-100",				"7.21e-100",				"3.7e-100",				"5.64e-100",				"7.7e-100",				"-8.03e-100",				"7.79e-100",				"7.98e-100",				"-2.71e-100",				"2.99e-100",				"6.78e-100",				"-7.99e-100",				"3.03e-100",				"-3.22e-100",				"2.76e-100",				"5.83e-100",				"-9.45e-100",				"-6.37e-100",				"-7.84e-100",				"9.94e-100",				"-2.6e-100",				"-8.29e-100",				"2.76e-100",				"-1.21e-100",				"1.69e-100",				"9.69e-100",				"-3.93e-100",				"-4.51e-100",				"7.86e-100",				"3.2e-100",				"-5.13e-101",				"1.17e-100",				"4.28e-100",				"-9.44e-100",				"-1.79e-100",				"-6.95e-100",				"6.2e-100",				"-1.94e-100",				"-8.6e-100",				"-1.14e-100",				"8.19e-100",				"8.23e-100",				"8.32e-100",				"-2.06e-100",				"7.01e-100",				"-4.9e-100",				"4.28e-100",				"5.3e-100",				"-4.09e-100",				"2.49e-100",				"-5.59e-100",				"-8.53e-100",				"9.64e-100",				"6.61e-100",				"-8.08e-100",				"-5.14e-100",				"-3.69e-100",				"5.88e-100",				"-6.07e-100",				"-3.75e-101" 			], 			"description" : "",			"name" : "contNarrow",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"0.269038564",				"-0.751745664",				"-0.70719854",				"-3.315197368",				"0.452474276",				"0.484832885",				"0.246931163",				"1.454038184",				"0.314986582",				"1.254507244",				"-1.056739214",				"0.027473107",				"2.010655929",				"-0.473325284",				"-0.218319199",				"-0.362584916",				"0.133472482",				"1.301359718",				"4.5",				"0.02063217",				"-0.660865856",				"0.813252697",				"0.453708481",				"1.142966476",				"0.524149442",				"-0.795363656",				"-1.138422",				"0.336652361",				"0.209776766",				"-0.86190895",				"1.856756889",				"-0.966375392",				"-0.440281387",				"-23",				"1.227390828",				"-0.266525133",				"-1.0809194",				"0.817291146",				"2.856221513",				"1.371816795",				"-3.12",				"1.569206836",				"0.104621969",				"-1.024937933",				"0.658332312",				"-0.72708755",				"0.01846117",				"-1.329301737",				"0.202589917",				"-0.539687644",				"-0.05383627",				"-0.011038078",				"-0.144103418",				"-0.118616735",				"0.435547301",				"-0.440890933",				"-1.602644626",				"0.488347655",				"5.7",				"-0.199209337",				"0.734678153",				"0.191952071",				"0.951021036",				"-0.12373379",				"-1.111496413",				"0.795501903",				"-0.05334565",				"-1.520094148",				"1.5008575",				"-0.337265933",				"0.087851546",				"-0.539954703",				"1.120564996",				"-0.547097953",				"1.184352504",				"-0.424653255",				"2.797608158",				"1.333281907",				"-0.070558974",				"-1.314625747",				"-0.946906667",				"-0.87197756",				"12",				"-1.589177705",				"-1.08767037",				"1.9003493",				"0.497966746",				"0.314781235",				"1.783908111",				"-0.496077535",				"-0.968971249",				"-0.918706246",				"-0.259830143",				"0.597605669",				"0.930721996",				"0.522670913",				"-0.921371121",				"-0.725693462",				"0.124421245",				"1.162498447" 			], 			"description" : "",			"name" : "contOutlier",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1.521033162",				"-0.910386819",				"-0.612928686",				"-1.332980719",				"-0.027805872",				"1.276411515",				"0.145010456",				"-0.964470142",				"2.006165806",				"0.86064292",				"-0.828902249",				"-0.611646545",				"1.858228623",				"0.642043149",				"-0.913900494",				"0.348716082",				"0.976441678",				"-0.526381444",				"0.332000711",				"-0.106530923",				"-1.002178797",				"0.626378919",				"1.484293596",				"0.374450749",				"1.270672337",				"-2.073133408",				"-1.058258102",				"0.421280454",				"-1.356023197",				"0.648830867",				"-0.082624118",				"-0.425815213",				"0.393831644",				"1.836337366",				"0.277718896",				"-0.833115983",				"1.055700614",				"-1.585812858",				"0.03507676",				"-0.397021392",				"1.43939299",				"-0.148337753",				"-1.355751095",				"-2.147231812",				"-0.560945149",				"0.612323493",				"-0.432126651",				"-0.265197642",				"0.299985335",				"0.825507709",				"0.132982528",				"-0.535487932",				"-0.758838087",				"0.512946962",				"2.361887648",				"1.277235754",				"0.963233262",				"-0.92409465",				"-0.328788884",				"-0.614385435",				"0.882268499",				"-0.61508922",				"-0.233919845",				"0.125705487",				"0.75460153",				"-0.916583292",				"-0.838739629",				"-1.044104412",				"1.835611751",				"1.257452217",				"-1.004617321",				"-0.167033164",				"-0.246003096",				"0.440896777",				"0.399083534",				"1.085484471",				"-0.186281719",				"0.831242032",				"-0.337926213",				"-0.035583912",				"-0.876635575",				"0.242603932",				"0.475305793",				"1.922957002",				"-0.864830262",				"-1.50569819",				"0.290460747",				"-1.151350671",				"1.06322576",				"0.645660111",				"2.263460985",				"-1.339377181",				"-1.036964374",				"0.793807358",				"-1.508132028",				"-0.507807907",				"0.361178166",				"0.817098677",				"-1.345254585",				"1.43303512" 			], 			"description" : "",			"name" : "contcor1",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1.480950383",				"-1.013220053",				"-0.296117034",				"0.089105927",				"0.517458031",				"0.630621107",				"0.338049419",				"-1.540254853",				"1.370681986",				"1.330696219",				"-0.102756392",				"0.424124199",				"2.024579557",				"0.171809484",				"0.063506238",				"1.57988411",				"0.905650485",				"-0.728247624",				"0.590554201",				"-0.336862124",				"-0.636844309",				"1.831881884",				"2.134815938",				"1.026428056",				"1.285324181",				"-0.768743297",				"-1.489765825",				"-0.01636327",				"-1.257525134",				"-0.014202675",				"-0.068970666",				"0.510773233",				"0.104634243",				"0.226108977",				"-0.155389747",				"-1.127839217",				"0.45790074",				"-1.217982555",				"-0.407814327",				"-0.298569173",				"0.768509543",				"0.874278791",				"-0.520476438",				"-1.729626099",				"0.399656151",				"0.714870906",				"-0.906914708",				"-0.692746481",				"-0.188711694",				"0.967936613",				"0.112492177",				"0.194084995",				"-1.10097767",				"1.076899229",				"1.458690501",				"0.464273155",				"1.892438271",				"-1.107328768",				"0.765559844",				"-0.468164112",				"-0.051912676",				"-0.255179814",				"1.765082916",				"-0.716611431",				"-0.386964684",				"-0.700729698",				"-0.327938549",				"-0.893630608",				"2.130133368",				"0.465795486",				"-0.753844325",				"-0.499094509",				"0.140877166",				"-0.003659941",				"-1.333467086",				"0.184361022",				"1.191267816",				"0.813489745",				"-0.413833871",				"1.13960029",				"-0.394914455",				"-2.709936005",				"0.777086628",				"-0.166163139",				"0.11573855",				"-1.558703131",				"-0.004748574",				"-1.094781613",				"1.255144556",				"-0.097987678",				"1.955082542",				"0.457211285",				"-2.417509866",				"1.274912805",				"-0.354674221",				"-1.066068935",				"-1.327338048",				"0.127761789",				"-0.456583351",				"0.568722799" 			], 			"description" : "",			"name" : "contcor2",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"m",				"f",				"m",				"f",				"m",				"m",				"f",				"m",				"f",				"f",				"f",				"m",				"f",				"m",				"m",				"m",				"m",				"f",				"m",				"m",				"m",				"f",				"m",				"f",				"m",				"m",				"f",				"f",				"f",				"m",				"m",				"f",				"f",				"f",				"f",				"m",				"f",				"m",				"m",				"m",				"f",				"m",				"f",				"m",				"f",				"f",				"f",				"m",				"f",				"m",				"f",				"m",				"m",				"f",				"m",				"f",				"f",				"m",				"f",				"m",				"m",				"f",				"f",				"m",				"m",				"f",				"m",				"f",				"f",				"m",				"f",				"m",				"f",				"m",				"f",				"m",				"f",				"m",				"m",				"m",				"m",				"m",				"f",				"m",				"f",				"f",				"m",				"m",				"m",				"f",				"f",				"f",				"f",				"f",				"f",				"m",				"f",				"f",				"m",				"f" 			], 			"description" : "",			"name" : "facGender",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental",				"control",				"experimental" 			], 			"description" : "",			"name" : "facExperim",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5",				"1",				"2",				"3",				"4",				"5" 			], 			"description" : "",			"name" : "facFive",			"rCode" : "",			"title" : "",			"type" : "ordinal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1",				"2",				"3",				"4",				"5",				"6",				"7",				"8",				"9",				"10",				"11",				"12",				"13",				"14",				"15",				"16",				"17",				"18",				"19",				"20",				"21",				"22",				"23",				"24",				"25",				"26",				"27",				"28",				"29",				"30",				"31",				"32",				"33",				"34",				"35",				"36",				"37",				"38",				"39",				"40",				"41",				"42",				"43",				"44",				"45",				"46",				"47",				"48",				"49",				"50",				"1",				"2",				"3",				"4",				"5",				"6",				"7",				"8",				"9",				"10",				"11",				"12",				"13",				"14",				"15",				"16",				"17",				"18",				"19",				"20",				"21",				"22",				"23",				"24",				"25",				"26",				"27",				"28",				"29",				"30",				"31",				"32",				"33",				"34",				"35",				"36",				"37",				"38",				"39",				"40",				"41",				"42",				"43",				"44",				"45",				"46",				"47",				"48",				"49",				"50" 			], 			"description" : "",			"name" : "facFifty",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f1",				"f2",				"f3",				"totallyridiculoussuperlongfactorme" 			], 			"description" : "",			"name" : "facOutlier",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"h",				"l",				"o",				"m",				"z",				"x",				"n",				"v",				"r",				"q",				"z",				"e",				"p",				"g",				"i",				"a",				"c",				"p",				"y",				"j",				"s",				"k",				"s",				"z",				"m",				"x",				"r",				"g",				"a",				"f",				"m",				"w",				"o",				"z",				"a",				"e",				"x",				"t",				"d",				"g",				"i",				"l",				"u",				"h",				"j",				"d",				"f",				"e",				"q",				"o",				"q",				"z",				"c",				"d",				"c",				"m",				"d",				"j",				"y",				"g",				"f",				"b",				"r",				"w",				"t",				"s",				"w",				"m",				"o",				"x",				"z",				"m",				"h",				"q",				"j",				"a",				"m",				"f",				"t",				"c",				"t",				"d",				"b",				"f",				"b",				"i",				"o",				"f",				"d",				"i",				"z",				"e",				"x",				"a",				"z",				"f",				"u",				"z",				"r",				"s" 			], 			"description" : "",			"name" : "debString",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"12.29505445",				"15.81197434",				"2.213892619",				"-9.146498869",				"14.17165658",				"-5.247577291",				"34.91974603",				"-11.49774025",				"16.27130461",				"-26.32751885",				"22.51349443",				"-22.29764163",				"2.079802765",				"-23.85909038",				"-15.10502073",				"58.3179159",				"-29.50332183",				"18.62680303",				"-34.4206158",				"8.60973483",				"-18.56055899",				"2.276286087",				"-7.071277306",				"-59.86347421",				"31.04606599",				"16.19107226",				"54.33924971",				"-4.974043519",				"-0.555140607",				"45.47605434",				"1.330889209",				"20.89165938",				"34.32134416",				"-15.30218115",				"-10.51401033",				"34.00566913",				"10.87473713",				"15.06146133",				"14.96051082",				"-40.08210082",				"8.218437573",				"-9.872269548",				"13.46850734",				"-20.01871895",				"16.38070237",				"-6.508017194",				"21.73765541",				"18.83402554",				"-39.15662657",				"22.31499477",				"-7.964379196",				"8.850454495",				"21.78692693",				"-24.76457867",				"-16.98676191",				"-32.34628474",				"0.451974986",				"-1.984585065",				"8.052450308",				"-9.245006147",				"-13.79539297",				"14.90191342",				"40.55177378",				"37.27318614",				"-11.1602374",				"40.89152002",				"23.25180016",				"",				"35.52541498",				"9.557958862",				"41.20062693",				"24.67674538",				"1.061635707",				"-8.666785017",				"-16.22418308",				"28.56596845",				"58.3438749",				"25.17760652",				"-40.14831628",				"30.55579311",				"-49.21614118",				"39.49936214",				"54.04293202",				"6.362484276",				"23.68813974",				"-21.46913233",				"-7.895855523",				"43.29461563",				"-23.70390365",				"-36.82168652",				"-56.19342872",				"14.40925333",				"5.212835789",				"-13.17381705",				"-36.23515279",				"-48.99978309",				"2.878536633",				"-14.26476856",				"40.20161446",				"3.309375769" 			], 			"description" : "",			"name" : "debMiss1",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"-12.50897286",				"",				"-6.105561134",				"-0.81577699",				"",				"25.74968991",				"9.318638024",				"-37.34994081",				"-26.24251961",				"64.50668921",				"",				"5.185419296",				"0.490318741",				"-1.708067004",				"-3.724596596",				"31.93803283",				"",				"",				"",				"3.171606677",				"",				"-26.03093825",				"45.18686658",				"8.83853789",				"12.71106193",				"",				"-17.99711621",				"",				"18.40487547",				"-3.166165593",				"16.22066096",				"",				"37.15164556",				"-5.230861965",				"20.6970882",				"11.63175636",				"20.93077096",				"-9.456293275",				"12.23958598",				"",				"7.192798371",				"-8.749054106",				"",				"",				"",				"-12.76054352",				"-4.580751279",				"9.135909664",				"53.11149213",				"",				"55.8084898",				"19.40455001",				"",				"15.80012906",				"14.80356943",				"8.247581677",				"-38.22842921",				"21.37372726",				"",				"-4.16753873",				"",				"21.74813188",				"",				"25.43540058",				"4.292728936",				"50.93755334",				"-7.708338532",				"23.94322819",				"37.88150118",				"2.621664702",				"15.76586814",				"",				"33.32318658",				"-32.93533803",				"-37.27169159",				"",				"12.79755723",				"38.54814123",				"25.13557785",				"",				"-15.59242229",				"",				"-27.14959186",				"-23.45857139",				"",				"",				"4.844530415",				"",				"",				"16.89675221",				"",				"",				"44.97152935",				"",				"-3.027361501",				"16.68448256",				"20.0129414",				"-40.50763657",				"43.54728062",				"-21.40626084" 			], 			"description" : "",			"name" : "debMiss30",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"",				"",				"-27.5946802",				"",				"",				"",				"",				"34.97482541",				"",				"",				"",				"-9.039264777",				"",				"",				"",				"",				"",				"43.63876663",				"59.98295451",				"",				"",				"",				"",				"20.18332567",				"-21.97367833",				"",				"",				"",				"",				"",				"9.750882176",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"-9.424529551",				"",				"",				"",				"",				"2.329880838",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"-21.05597816",				"-5.761227735",				"",				"",				"31.63156258",				"",				"",				"",				"-16.87864345",				"",				"34.91209272",				"4.100531341",				"",				"",				"",				"",				"",				"",				"",				"25.39075658",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"18.95203462",				"-22.60679192",				"36.00956513",				"",				"",				"",				"",				"",				"",				"" 			], 			"description" : "",			"name" : "debMiss80",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"24.1187081",				"",				"",				"",				"",				"",				"",				"" 			], 			"description" : "",			"name" : "debMiss99",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1",				"",				"0",				"0",				"0",				"0",				"1",				"0",				"0",				"1",				"0",				"0",				"1",				"1",				"1",				"1",				"1",				"0",				"1",				"0",				"0",				"0",				"0",				"1",				"",				"",				"",				"",				"0",				"",				"0",				"0",				"1",				"",				"1",				"1",				"0",				"1",				"1",				"1",				"1",				"0",				"1",				"0",				"",				"0",				"0",				"",				"1",				"0",				"0",				"",				"0",				"1",				"1",				"1",				"1",				"1",				"1",				"",				"",				"0",				"0",				"1",				"",				"",				"1",				"",				"1",				"0",				"",				"1",				"1",				"0",				"1",				"1",				"1",				"1",				"1",				"1",				"1",				"0",				"1",				"0",				"0",				"0",				"1",				"1",				"0",				"0",				"",				"",				"1",				"1",				"1",				"",				"1",				"1",				"0",				"" 			], 			"description" : "",			"name" : "debBinMiss20",			"rCode" : "",			"title" : "",			"type" : "nominal" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"",				"" 			], 			"description" : "",			"name" : "debNaN",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"19.33621882",				"-19.8459184",				"-15.38229892",				"-14.57178797",				"",				"-0.538542085",				"7.379238519",				"47.63552969",				"41.06299019",				"7.391813158",				"12.56673523",				"40.18473672",				"7.186747009",				"17.26707742",				"13.43839211",				"-6.017744385",				"-10.5254181",				"60.33055091",				"-33.17939265",				"42.13211",				"41.63230666",				"18.08252447",				"23.19866523",				"-4.140889402",				"-2.585273933",				"-2.452483948",				"2.645337056",				"27.43954953",				"",				"-4.010210335",				"-14.27635326",				"-6.324200517",				"-52.57099368",				"14.9125812",				"-23.5585338",				"9.777477989",				"27.81169355",				"-9.855238762",				"56.14853606",				"-1.074534587",				"-21.10145059",				"14.2327728",				"7.749639899",				"64.30898882",				"8.578501694",				"27.56515752",				"13.10111222",				"",				"55.13509021",				"11.36083995",				"24.14384346",				"-11.84490983",				"17.55972289",				"5.154103894",				"-16.43481581",				"11.79509032",				"25.67878301",				"-26.56105033",				"30.89992274",				"-7.803536699",				"26.07333913",				"0.047147038",				"-25.21252526",				"-10.13829937",				"",				"1.945479714",				"2.40498707",				"26.67366838",				"",				"9.211564035",				"",				"33.32843989",				"48.54918845",				"31.83812445",				"16.92271037",				"6.075765038",				"24.82786846",				"35.73968397",				"21.48868763",				"38.94721253",				"51.92014548",				"24.98712874",				"-36.71891581",				"",				"21.12029096",				"-11.41658279",				"43.90869498",				"7.78832157",				"5.924135526",				"45.41756515",				"",				"",				"8.333502869",				"40.0390802",				"19.71146144",				"27.47115607",				"-3.416596104",				"",				"28.32081039",				"32.97384623" 			], 			"description" : "",			"name" : "debNaN10",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e",				"\u221e" 			], 			"description" : "",			"name" : "debInf",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"0.642038201",				"0.515440275",				"0.447585321",				"0.642930641",				"0.585416105",				"0.544931008",				"0.743580248",				"0.651916557",				"0.696833552",				"0.694601235",				"0.711955085",				"0.593240305",				"0.694789138",				"0.736839036",				"0.698094646",				"0.526075461",				"0.633042095",				"0.671483094",				"0.644589319",				"0.638244615",				"0.869855592",				"0.801496435",				"0.704276279",				"0.666621926",				"0.690001843",				"0.563699434",				"0.698381203",				"0.590536511",				"0.6716216",				"0.634324908",				"0.7115023",				"0.579010829",				"0.603432164",				"0.685082521",				"0.652887496",				"0.581590324",				"0.675894708",				"0.665411145",				"0.757228652",				"0.62041137",				"0.650746613",				"0.679484841",				"0.599784155",				"0.673688173",				"0.767064874",				"0.584264788",				"0.572718084",				"0.659782958",				"0.661300458",				"0.617465311",				"0.783959893",				"0.706065325",				"0.625826063",				"0.574662184",				"0.738471737",				"0.754113413",				"0.684338249",				"0.763499072",				"0.843919117",				"0.72013878",				"0.733252493",				"0.699404524",				"0.687269622",				"0.565707078",				"0.688885545",				"0.673945048",				"0.721324295",				"0.561952238",				"0.487838809",				"0.531128923",				"0.687233474",				"0.678345322",				"0.565085458",				"0.682553643",				"0.507997006",				"0.761748559",				"0.758175022",				"0.747018421",				"0.629427634",				"0.694269237",				"0.678784917",				"0.600199885",				"0.56663707",				"0.699231378",				"0.702445078",				"0.799151871",				"0.68889492",				"0.550942502",				"0.654394642",				"0.636740024",				"0.570627192",				"0.495749838",				"0.757810004",				"0.82636797",				"0.75677294",				"0.662298439",				"0.602703674",				"0.807596206",				"0.584414672",				"0.690641762" 			], 			"description" : "",			"name" : "debCollin1",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"2.642038201",				"2.515440275",				"2.447585321",				"2.642930641",				"2.585416105",				"2.544931008",				"2.743580248",				"2.651916557",				"2.696833552",				"2.694601235",				"2.711955085",				"2.593240305",				"2.694789138",				"2.736839036",				"2.698094646",				"2.526075461",				"2.633042095",				"2.671483094",				"2.644589319",				"2.638244615",				"2.869855592",				"2.801496435",				"2.704276279",				"2.666621926",				"2.690001843",				"2.563699434",				"2.698381203",				"2.590536511",				"2.6716216",				"2.634324908",				"2.7115023",				"2.579010829",				"2.603432164",				"2.685082521",				"2.652887496",				"2.581590324",				"2.675894708",				"2.665411145",				"2.757228652",				"2.62041137",				"2.650746613",				"2.679484841",				"2.599784155",				"2.673688173",				"2.767064874",				"2.584264788",				"2.572718084",				"2.659782958",				"2.661300458",				"2.617465311",				"2.783959893",				"2.706065325",				"2.625826063",				"2.574662184",				"2.738471737",				"2.754113413",				"2.684338249",				"2.763499072",				"2.843919117",				"2.72013878",				"2.733252493",				"2.699404524",				"2.687269622",				"2.565707078",				"2.688885545",				"2.673945048",				"2.721324295",				"2.561952238",				"2.487838809",				"2.531128923",				"2.687233474",				"2.678345322",				"2.565085458",				"2.682553643",				"2.507997006",				"2.761748559",				"2.758175022",				"2.747018421",				"2.629427634",				"2.694269237",				"2.678784917",				"2.600199885",				"2.56663707",				"2.699231378",				"2.702445078",				"2.799151871",				"2.68889492",				"2.550942502",				"2.654394642",				"2.636740024",				"2.570627192",				"2.495749838",				"2.757810004",				"2.82636797",				"2.75677294",				"2.662298439",				"2.602703674",				"2.807596206",				"2.584414672",				"2.690641762" 			], 			"description" : "",			"name" : "debCollin2",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"1.284076402",				"1.03088055",				"0.895170641",				"1.285861281",				"1.170832209",				"1.089862015",				"1.487160496",				"1.303833114",				"1.393667105",				"1.389202469",				"1.42391017",				"1.186480609",				"1.389578277",				"1.473678072",				"1.396189293",				"1.052150922",				"1.26608419",				"1.342966187",				"1.289178637",				"1.27648923",				"1.739711184",				"1.60299287",				"1.408552559",				"1.333243853",				"1.380003687",				"1.127398868",				"1.396762406",				"1.181073022",				"1.343243199",				"1.268649816",				"1.423004601",				"1.158021658",				"1.206864328",				"1.370165042",				"1.305774992",				"1.163180647",				"1.351789417",				"1.33082229",				"1.514457304",				"1.24082274",				"1.301493226",				"1.358969682",				"1.199568311",				"1.347376346",				"1.534129747",				"1.168529577",				"1.145436168",				"1.319565916",				"1.322600917",				"1.234930622",				"1.567919787",				"1.41213065",				"1.251652125",				"1.149324368",				"1.476943475",				"1.508226826",				"1.368676499",				"1.526998143",				"1.687838235",				"1.440277559",				"1.466504987",				"1.398809047",				"1.374539244",				"1.131414156",				"1.37777109",				"1.347890096",				"1.442648591",				"1.123904475",				"0.975677618",				"1.062257846",				"1.374466948",				"1.356690644",				"1.130170917",				"1.365107287",				"1.015994012",				"1.523497119",				"1.516350044",				"1.494036842",				"1.258855267",				"1.388538473",				"1.357569834",				"1.20039977",				"1.133274141",				"1.398462756",				"1.404890156",				"1.598303742",				"1.377789839",				"1.101885004",				"1.308789284",				"1.273480047",				"1.141254385",				"0.991499677",				"1.515620007",				"1.652735939",				"1.513545881",				"1.324596877",				"1.205407348",				"1.615192412",				"1.168829344",				"1.381283524" 			], 			"description" : "",			"name" : "debCollin3",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"12.59439792",				"3.203714829",				"2.039632761",				"23.28135794",				"11.94984212",				"5.86162875",				"1.860119237",				"2.172089504",				"4.794362541",				"2.772569488",				"29.32237861",				"3.289304957",				"29.60816458",				"12.81628158",				"10.33111466",				"7.339607773",				"14.74490268",				"19.04092749",				"4.457849079",				"0.981495504",				"18.86640292",				"7.09200338",				"36.20041863",				"10.76760955",				"11.75533742",				"1.470873667",				"0.058757922",				"3.088944846",				"3.702698217",				"8.499626141",				"1.481335791",				"12.29630114",				"10.31847212",				"0.921557925",				"0.817209248",				"1.75851692",				"1.302255697",				"60.62183473",				"5.902888641",				"11.62692353",				"4.086470137",				"10.21653027",				"5.023304044",				"16.0212847",				"6.029500724",				"7.594024222",				"6.529850511",				"35.44104618",				"3.763125443",				"0.227627757",				"2.532787387",				"0.800762803",				"10.4416378",				"2.735722696",				"20.92707249",				"7.41450701",				"0.197014145",				"8.010464244",				"3.893310436",				"9.851298775",				"27.54739748",				"24.8250967",				"11.10667539",				"1.720398897",				"21.90849398",				"9.618616708",				"8.150790987",				"48.06656736",				"5.430236147",				"58.16481289",				"5.466777644",				"0.211640329",				"4.725582498",				"9.065711464",				"20.76629463",				"25.86267557",				"2.230258842",				"17.35546794",				"14.00530378",				"1.853272735",				"16.34066342",				"53.17142645",				"7.493237606",				"2.078883849",				"3.16479117",				"6.252009424",				"8.378574001",				"13.73741003",				"13.20115916",				"1.613181883",				"5.389407967",				"8.377217386",				"1.275822202",				"25.99118845",				"18.38902919",				"9.893418768",				"2.66166856",				"6.540878081",				"0.43232324",				"6.381575258" 			], 			"description" : "",			"name" : "debEqual1",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"12.59439792",				"3.203714829",				"2.039632761",				"23.28135794",				"11.94984212",				"5.86162875",				"1.860119237",				"2.172089504",				"4.794362541",				"2.772569488",				"29.32237861",				"3.289304957",				"29.60816458",				"12.81628158",				"10.33111466",				"7.339607773",				"14.74490268",				"19.04092749",				"4.457849079",				"0.981495504",				"18.86640292",				"7.09200338",				"36.20041863",				"10.76760955",				"11.75533742",				"1.470873667",				"0.058757922",				"3.088944846",				"3.702698217",				"8.499626141",				"1.481335791",				"12.29630114",				"10.31847212",				"0.921557925",				"0.817209248",				"1.75851692",				"1.302255697",				"60.62183473",				"5.902888641",				"11.62692353",				"4.086470137",				"10.21653027",				"5.023304044",				"16.0212847",				"6.029500724",				"7.594024222",				"6.529850511",				"35.44104618",				"3.763125443",				"0.227627757",				"2.532787387",				"0.800762803",				"10.4416378",				"2.735722696",				"20.92707249",				"7.41450701",				"0.197014145",				"8.010464244",				"3.893310436",				"9.851298775",				"27.54739748",				"24.8250967",				"11.10667539",				"1.720398897",				"21.90849398",				"9.618616708",				"8.150790987",				"48.06656736",				"5.430236147",				"58.16481289",				"5.466777644",				"0.211640329",				"4.725582498",				"9.065711464",				"20.76629463",				"25.86267557",				"2.230258842",				"17.35546794",				"14.00530378",				"1.853272735",				"16.34066342",				"53.17142645",				"7.493237606",				"2.078883849",				"3.16479117",				"6.252009424",				"8.378574001",				"13.73741003",				"13.20115916",				"1.613181883",				"5.389407967",				"8.377217386",				"1.275822202",				"25.99118845",				"18.38902919",				"9.893418768",				"2.66166856",				"6.540878081",				"0.43232324",				"6.381575258" 			], 			"description" : "",			"name" : "debEqual2",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3",				"12.3" 			], 			"description" : "",			"name" : "debSame",			"rCode" : "",			"title" : "",			"type" : "scale" 		}, 		{			"autoSortByValue" : true,			"codeType" : "notComputed",			"constructorJson" : {} ,			"customEmptyValues" : 			{				"custom" : false,				"strings" : [] 			}, 			"data" : 			[				"\u00a4_\u03b1_\u0430_\u05d0_\u1200_\u0bd0_\u0950_\u16a0_\u1100",				"\u00a2_\ud835\udec2_\u25cc\u2df6_\u2135_\u1201_\u0b85_\ua8fd_\u16a1_\u3131",				"\uffe0_\ud835\udefc_\u0410_\ufb21_\u1202_\u0b86_\u0972_\u16a2_\u3260",				"$_\ud835\udf36_\u04d1_\u05d0\u05b7_\u1203_\u0b87_\u0904_\u16a4_\uffa1",				"\uff04_\ud835\udf70_\u04d0_\u05d0\u05b8_\u1204_\u0b88_\u0905_\u16a5_\uac00",				"\ufe69_\ud835\udfaa_\u04d3_\u05d0\u05bc_\u1205_\u0b89_\u0906_\u16a6_\uac00",				"\u00a3_\u0391_\u04d2_\ufb4f_\u1206_\u0b8a_\u0973_\u16a7_\u326e",				"\uffe1_\ud835\udea8_\u04d9_\u05d1_\u1207_\u0b8e_\u0974_\u16f0_\uac00\u1161",				"\u00a5_\ud835\udee2_\u04d8_\u2136_\u1208_\u0b8f_\u0975_\u16a8_\uac00\u1175",				"\uffe5_\ud835\udf1c_\u04db_\u05d1\u05bc_\u1209_\u0b90_\u0976_\u16a9_\uac01",				"\u058f_\ud835\udf56_\u04da_\u05d1\u05bf_\u120a_\u0b92_\u0977_\u16ac_\uac02",				"\u060b_\ud835\udf90_\u04d5_\u05d2_\u120b_\u0b93_\u0907_\u16ad_\uac03",				"\u07fe_\u1f00_\u04d4_\u2137_\u120c_\u0b94_\u0908_\u16ae_\uac04",				"\u07ff_\u1f08_\u0431_\u05d2\u05bc_\u120d_\u0b94_\u0909_\u16f4_\uac05",				"\u09f2_\u1f04_\u25cc\u2de0_\u05d3_\u120e_\u0b83_\u090a_\u16af_\uae30",				"\u09f3_\u1f0c_\u0411_\u2138_\u120f_\u0b95_\u090b_\u16b0_\u1101",				"\u09fb_\u1f84_\u0432_\ufb22_\u2d80_\u0b99_\u0960_\u16b1_\u3132",				"\u0af1_\u1f8c_\u1c80_\u05d3\u05bc_\u1210_\u0b9a_\u090c_\u16b2_\uffa2",				"\ua838_\u1f02_\u25cc\u2de1_\u05d4_\u1211_\u0b9e_\u0961_\u16b3_\u1102",				"\u0bf9_\u1f0a_\u0412_\ufb23_\u1212_\u0b9f_\u090d_\u16b4_\u3134",				"\ud807\udfdd_\u1f82_\u0433_\u05d4\u05bc_\u1213_\u0ba3_\u090e_\u16b5_\u3261",				"\ud807\udfde_\u1f8a_\u25cc\u2de2_\u05d5_\u1214_\u0ba4_\u090f_\u16b6_\uffa4",				"\ud807\udfdf_\u1f06_\u0413_\u05d5\u05b9_\u1215_\u0ba8_\u0910_\u16f1_\u326f",				"\ud807\udfe0_\u1f0e_\u0453_\u05d5\u05bc_\u1216_\u0baa_\ua8fe_\u16b7_\u1103",				"\u0e3f_\u1f86_\u0403_\u05f0_\u1217_\u0bae_\u0911_\u16b9_\u3137",				"\u17db_\u1f8e_\u0491_\u05f1_\u1218_\u0baf_\u0912_\u16e9_\u3262",				"\ud838\udeff_\u1f80_\u0490_\u05d6_\u1219_\u0bb0_\u0913_\u16ba_\uffa7",				"\u20a0_\u1f88_\u0493_\u05d6\u05bc_\u121a_\u0bb2_\u0914_\u16bb_\u3270",				"\u20a1_\u1f01_\u0492_\u05d7_\u121b_\u0bb5_\u0915_\u16bc_\u1104",				"\u20a2_\u1f09_\u04fb_\u05d8_\u121c_\u0bb4_\u0915\u093c_\u16bd_\u3138",				"\u20a3_\u1f05_\u04fa_\u05d8\u05bc_\u121d_\u0bb3_\u0916_\u16be_\uffa8",				"\u20a4_\u1f0d_\u0495_\u05d9_\u121e_\u0bb1_\u0916\u093c_\u16bf_\u1105",				"\u20a5_\u1f85_\u0494_\u05d9\u05b4_\u121f_\u0ba9_\u0917_\u16c0_\u3139",				"\u20a6_\u1f8d_\u04f7_\u05d9\u05bc_\u1380_\u0b9c_\u0917\u093c_\u16c1_\u3263",				"\u20a7_\u1f03_\u04f6_\u05ef_\u1381_\u0bb6_\u097b_\u16c2_\uffa9",				"\u20a9_\u1f0b_\u0434_\u05f2_\u1382_\u0bb7_\u0918_\u16f5_\u3271",				"\uffe6_\u1f83_\u1c81_\u05f2\u05b7_\u1383_\u0bb8_\u0919_\u16c3_\u1106",				"\u20aa_\u1f8b_\u25cc\u2de3_\u05db_\u2d81_\u0bb9_\u091a_\u16c4_\u3141",				"\u20ab_\u1f07_\u0414_\ufb24_\u1220_\u25cc\u0bbe_\u091b_\u16c5_\u3264",				"\u20ac_\u1f0f_\u0501_\u05da_\u1221_\u25cc\u0bbf_\u091c_\u16c6_\uffb1",				"\u20ad_\u1f87_\u0500_\u05db\u05bc_\u1222_\u25cc\u0bc0_\u091c\u093c_\u16ee_\u3272",				"\u20ae_\u1f8f_\ua681_\u05da\u05bc_\u1223_\u25cc\u0bc1_\u0979_\u16c7_\u1107",				"\u20af_\u1f81_\ua680_\u05db\u05bf_\u1224_\u25cc\u0bc2_\u097c_\u16c8_\u3142",				"\u20b0_\u1f89_\u0452_\u05dc_\u1225_\u25cc\u0bc6_\u091d_\u16d5_\u3265",				"\u20b1_\u03ac_\u0402_\ufb25_\u1226_\u25cc\u0bc7_\u091e_\u16c9_\uffb2",				"\u20b2_\u03ac_\ua663_\u05dc\u05bc_\u1227_\u25cc\u0bc8_\u091f_\u16ca_\u3273",				"\u20b3_\u0386_\ua662_\u05de_\u1228_\u25cc\u0bca_\u0920_\u16cb_\u1108",				"\u20b4_\u0386_\u0503_\ufb26_\u1229_\u25cc\u0bca_\u0978_\u16ea_\u3143",				"\u20b5_\u1fb4_\u0502_\u05dd_\u122a_\u25cc\u0bcb_\u0921_\u16cc_\uffb3",				"\u20b6_\u1f70_\u0499_\u05de\u05bc_\u122b_\u25cc\u0bcb_\u0921\u093c_\u16cd_\u1109",				"\u20b7_\u1fba_\u0498_\u05e0_\u122c_\u25cc\u0bcc_\u097e_\u16ce_\u3145",				"\u20b8_\u1fb2_\u0435_\u05df_\u122d_\u25cc\u0bcc_\u0922_\u16f2_\u3266",				"\u20b9_\u1fb0_\u25cc\u2df7_\u05e0\u05bc_\u122e_\u25cc\u0bcd_\u0922\u093c_\u16cf_\uffb5",				"\ud83b\udcb0_\u1fb8_\u0415_\u05e1_\u122f_\u25cc\u0bd7_\u0923_\u16d0_\u3274",				"\u20ba_\u1fb6_\u0450_\u05e1\u05bc_\u2d82__\u0924_\u16d1_\u110a",				"\u20bb_\u1fb7_\u0400_\u05e2_\u1230__\u0925_\u16d2_\u3146",				"\u20bc_\u1fb1_\u04d7_\ufb20_\u1231__\u0926_\u16d3_\uffb6",				"\u20bd_\u1fb9_\u04d6_\u05e4_\u1232__\u0927_\u16d4_\u110b",				"\u20be_\u1fb3_\u0451_\u05e3_\u1233__\u0928_\u16d6_\u3147",				"\u20bf_\u1fbc_\u0401_\u05e4\u05bc_\u1234__\u0929_\u16f6_\u3267",				"_\u03b2_\u0454_\u05e3\u05bc_\u1235__\u092a_\u16d7_\uffb7",				"_\u03d0_\u25cc\ua674_\u05e4\u05bf_\u1236__\u092b_\u16d8_\u3275",				"_\ud835\udec3_\u0404_\u05e6_\u1237__\u092b\u093c_\u16d9_\u327e",				"_\ud835\udefd_\u0436_\u05e5_\u2d83__\u092c_\u16ef_\u110c",				"_\ud835\udf37_\u25cc\u2de4_\u05e6\u05bc_\uab01__\u097f_\u16da_\u3148",				"_\ud835\udf71_\u0416_\u05e7_\uab02__\u092d_\u16db_\u3268",				"_\ud835\udfab_\u04c2_\u05e7\u05bc_\uab03__\u092e_\u16dc_\uffb8",				"_\u0392_\u04c1_\u05e8_\uab04__\u092f_\u16dd_\u3276",				"_\ud835\udea9_\u04dd_\ufb27_\uab05__\u092f\u093c_\u16de_\u327d",				"_\ud835\udee3_\u04dc_\u05e8\u05bc_\uab06__\u097a_\u16df_\u110d",				"_\ud835\udf1d_\u052b_\u05e9_\u1238__\u0930_\u16f3_\u3149",				"_\ud835\udf57_\u052a_\u05e9\u05c2_\u1239__\u0931_\u16aa_\uffb9",				"_\ud835\udf91_\ua685_\u05e9\u05c1_\u123a__\u0932_\u16f7_\u110e",				"_\u1d5d_\ua684_\u05e9\u05bc_\u123b__\u0933_\u16ab_\u314a",				"_\u1d66_\u0497_\u05e9\u05bc\u05c2_\u123c__\u0934_\u16f8_\u3269",				"_\u03b3_\u0496_\u05e9\u05bc\u05c1_\u123d__\u0935_\u16a3_\uffba",				"_\u213d_\u0437_\u05ea_\u123e__\u0936_\u16e0_\u3277",				"_\ud835\udec4_\u25cc\u2de5_\ufb28_\u123f__\u0937_\u16e3_\u327c",				"_\ud835\udefe_\u0417_\u05ea\u05bc_\u2d84__\u0938_\u16b8_\u110f",				"_\ud835\udf38_\u04df__\u1240__\u0939_\u16e4_\u314b",				"_\ud835\udf72_\u04de__\u1241__\u093d_\u16e1_\u326a",				"_\ud835\udfac_\ua641__\u1242__\u097d_\u16e2_\uffbb",				"_\u0393_\ua640__\u1243__\u1ce9_\u16e5_\u3278",				"_\u213e_\u0505__\u1244__\u1cea_\u16e6_\u1110",				"_\ud835\udeaa_\u0504__\u1245__\u1ceb_\u16e7_\u314c",				"_\ud835\udee4_\u0511__\u1246__\u1cec_\u16e8_\u326b",				"_\ud835\udf1e_\u0510__\u1247__\u1cee__\uffbc",				"_\ud835\udf58_\ua643__\u1248__\u1cef__\u3279",				"_\ud835\udf92_\ua642__\u124a__\u1cf0__\u1111",				"_\u1d5e_\u0455__\u124b__\u1cf1__\u314d",				"_\u1d67_\u0405__\u124c__\u1cfa__\u326c",				"_\u1d26_\ua645__\u124d__\u1cf5__\uffbd",				"_\u03b4_\ua644__\u1250__\u1cf6__\u327a",				"_\ud835\udec5_\u04e1__\u1251__\ua8f2__\u1112",				"_\ud835\udeff_\u04e0__\u1252__\ua8f3__\u314e",				"_\ud835\udf39_\ua689__\u1253__\ua8f4__\u326d",				"_\ud835\udf73_\ua688__\u1254__\ua8f5__\uffbe",				"_\ud835\udfad_\u0507__\u1255__\ua8f6__\ud558",				"_\u0394_\u0506__\u1256__\ua8f7__\u327b",				"_\ud835\udeab_\ua683__\u1258__\ua8fb__\ud788" 			], 			"description" : "",			"name" : "unicode",			"rCode" : "",			"title" : "",			"type" : "nominal" 		} 	], 	"customEmptyValues" : 	{		"custom" : true,		"strings" : 		[			".",			"NA",			"NaN",			"nan" 		] 	} })quotePower";
-
-Json::Value debugCsvJson()
-{
-	Json::Reader parser;
-	Json::Value  value;
-
-	parser.parse(___debugcsvtxt, value);
-
-	return value;
-}
-
-#endif // HARDCODEDDATA_H
